+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:go_router/go_router.dart';
 import 'package:isar/isar.dart';
 import 'package:url_strategy/url_strategy.dart';
 
-import 'firebase_options.dart';
+import 'firebase_options_export.dart';
 import 'router/router.dart';
 import 'theme/theme.dart';
 import 'theme/theme_provider.dart';