+++ lib/services/env.dart
@@ -1,0 +1,7 @@
+enum Env {
+  prod,
+  dev,
+}
+
+const envString = String.fromEnvironment('env', defaultValue: 'dev');
+final env = Env.values.byName(envString);