+++ lib/utils/value.dart
@@ -1,0 +1,10 @@
+class Value<T> {
+  const Value.absent()
+      : value = null,
+        hasValue = false;
+
+  const Value(this.value) : hasValue = true;
+
+  final T? value;
+  final bool hasValue;
+}